--- ScooterRental/Exceptions/InvalidIdException.cs
@@ -1,10 +1,0 @@
-﻿using System;
-
-namespace ScooterRental.Exceptions
-{
-    public class InvalidIdException : Exception
-    {
-        public InvalidIdException() :
-            base("ID cannot be null or empty") { }
-    }
-}
