--- ScooterRental/Exceptions/InvalidPriceException.cs
@@ -1,10 +1,0 @@
-﻿using System;
-
-namespace ScooterRental.Exceptions
-{
-    public class InvalidPriceException : Exception
-    {
-        public InvalidPriceException(decimal price) :
-            base($"Given price {price} not valid") { }
-    }
-}
