--- ScooterRental/Exceptions/ScooterAlreadyRentedException.cs
@@ -1,10 +1,0 @@
-﻿using System;
-
-namespace ScooterRental.Exceptions
-{
-    public class ScooterAlreadyRentedException : Exception
-    {
-        public ScooterAlreadyRentedException(string id) :
-            base($"Scooter with ID {id} already rented") { }
-    }
-}
