--- ScooterRental/Exceptions/ScooterDoesNotExistException.cs
@@ -1,10 +1,0 @@
-﻿using System;
-
-namespace ScooterRental.Exceptions
-{
-    public class ScooterDoesNotExistException : Exception
-    {
-        public ScooterDoesNotExistException(string id) :
-            base($"Scooter {id} does not exist") { }
-    }
-}
