+++ ScooterRentalAPI.Core/Calculators.cs
