+++ ScooterRentalAPI.Core/Exceptions/DuplicateScooterException.cs
