+++ ScooterRentalAPI.Core/Exceptions/InvalidIdException.cs
