+++ ScooterRentalAPI.Core/Exceptions/InvalidPriceException.cs
