+++ ScooterRentalAPI.Core/Exceptions/ScooterAlreadyRentedException.cs
