+++ ScooterRentalAPI.Core/Exceptions/ScooterDoesNotExistException.cs
