+++ ScooterRentalAPI.Core/Interfaces/IRentalCompany.cs
