+++ ScooterRentalAPI.Core/Interfaces/IScooterService.cs
