+++ ScooterRentalAPI.Core/Models/RentedScooter.cs
@@ -1,15 +1,15 @@
 ﻿using System;
 
-namespace ScooterRentalAPI
+namespace ScooterRentalAPI.Core.Models
 {
-    public class RentedScooters
+    public class RentedScooter
     {
         public int Id { get; set; }
         public DateTime StartTime { get; set; }
         public DateTime? EndTime { get; set; }
         public decimal PricePerMinute { get; set; }
 
-        public RentedScooters(int id, DateTime startTime, decimal pricePerMinute)
+        public RentedScooter(int id, DateTime startTime, decimal pricePerMinute)
         {
             Id = id;
             StartTime = startTime;
