+++ ScooterRentalAPI.Core/ScooterRentalAPI.Core.csproj
@@ -1,4 +1,4 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
+<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
     <TargetFramework>net5.0</TargetFramework>
@@ -9,12 +9,6 @@
       <PrivateAssets>all</PrivateAssets>
       <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
     </PackageReference>
-    <PackageReference Include="Swashbuckle.AspNetCore" Version="5.6.3" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\ScooterRentalAPI.Core\ScooterRentalAPI.Core.csproj" />
-    <ProjectReference Include="..\ScooterRentalAPI.Data\ScooterRentalAPI.Data.csproj" />
   </ItemGroup>
 
 </Project>
