+++ ScooterRentalAPI.Core/Validators/Validator.cs
@@ -1,12 +1,13 @@
-﻿using ScooterRental.Exceptions;
+﻿using ScooterRentalAPI.Core.Models;
+using ScooterRentalAPI.Exceptions;
 
-namespace ScooterRental.Validators
+namespace ScooterRentalAPI.Core.Validators
 {
     public static class Validator
     {
-        public static void ScooterIdValidator(string id)
+        public static void ScooterIdValidator(int id)
         {
-            if (string.IsNullOrEmpty(id))
+            if (id==null)
             {
                 throw new InvalidIdException();
             }
