+++ ScooterRentalAPI.sln
@@ -3,11 +3,13 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.0.31903.59
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ScooterRental", "ScooterRental\ScooterRental.csproj", "{C9F1EC6C-07D1-442E-9194-A15869D205F1}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ScooterRental.tests", "ScooterRental.tests\ScooterRental.tests.csproj", "{15231C45-6F91-457C-A2BD-0E2BF7BA0EC0}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ScooterRentalAPI", "ScooterRentalAPI\ScooterRentalAPI.csproj", "{ECCF9D91-534F-45EE-AE1B-16FC7A8DE316}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ScooterRentalAPI", "ScooterRentalAPI\ScooterRentalAPI.csproj", "{ECCF9D91-534F-45EE-AE1B-16FC7A8DE316}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ScooterRentalAPI.Data", "ScooterRentalAPI.Data\ScooterRentalAPI.Data.csproj", "{F7A290D0-922A-4F99-B830-3F100B8E4CA6}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ScooterRentalAPI.Core", "ScooterRentalAPI.Core\ScooterRentalAPI.Core.csproj", "{F3CD38E9-73BF-465D-B902-B2DE89BB5ED3}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,10 +17,6 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{C9F1EC6C-07D1-442E-9194-A15869D205F1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{C9F1EC6C-07D1-442E-9194-A15869D205F1}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{C9F1EC6C-07D1-442E-9194-A15869D205F1}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{C9F1EC6C-07D1-442E-9194-A15869D205F1}.Release|Any CPU.Build.0 = Release|Any CPU
 		{15231C45-6F91-457C-A2BD-0E2BF7BA0EC0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{15231C45-6F91-457C-A2BD-0E2BF7BA0EC0}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{15231C45-6F91-457C-A2BD-0E2BF7BA0EC0}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -27,6 +25,14 @@ Global
 		{ECCF9D91-534F-45EE-AE1B-16FC7A8DE316}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{ECCF9D91-534F-45EE-AE1B-16FC7A8DE316}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{ECCF9D91-534F-45EE-AE1B-16FC7A8DE316}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F7A290D0-922A-4F99-B830-3F100B8E4CA6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F7A290D0-922A-4F99-B830-3F100B8E4CA6}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F7A290D0-922A-4F99-B830-3F100B8E4CA6}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F7A290D0-922A-4F99-B830-3F100B8E4CA6}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F3CD38E9-73BF-465D-B902-B2DE89BB5ED3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F3CD38E9-73BF-465D-B902-B2DE89BB5ED3}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F3CD38E9-73BF-465D-B902-B2DE89BB5ED3}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F3CD38E9-73BF-465D-B902-B2DE89BB5ED3}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
