+++ ScooterRentalAPI/Startup.cs
@@ -1,16 +1,11 @@
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.HttpsPolicy;
-using Microsoft.AspNetCore.Mvc;
+using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
-using Microsoft.Extensions.Logging;
 using Microsoft.OpenApi.Models;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+using ScooterRentalAPI.Data;
 
 namespace ScooterRentalAPI
 {
@@ -32,6 +27,25 @@ namespace ScooterRentalAPI
             {
                 c.SwaggerDoc("v1", new OpenApiInfo { Title = "ScooterRentalAPI", Version = "v1" });
             });
+
+            services.AddDbContext<ScooterRentalDbContext>(options =>
+            {
+                options.UseSqlServer(Configuration.GetConnectionString("scooter-rental"));
+            });
+
+            services.AddScoped<ScooterRentalDbContext>();
+
+            //// Default Policy
+            //services.AddCors(options =>
+            //{
+            //    options.AddDefaultPolicy(
+            //        builder =>
+            //        {
+            //            builder.WithOrigins("https://localhost:7070", "http://localhost:4200")
+            //                .AllowAnyHeader()
+            //                .AllowAnyMethod();
+            //        });
+            //});
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
@@ -47,6 +61,8 @@ namespace ScooterRentalAPI
             app.UseHttpsRedirection();
 
             app.UseRouting();
+
+            //app.UseCors();
 
             app.UseAuthorization();
