+++ ScooterRentalAPI/appsettings.json
@@ -6,5 +6,8 @@
       "Microsoft.Hosting.Lifetime": "Information"
     }
   },
+  "ConnectionStrings": {
+    "scooter-rental": "Data Source=localhost;Initial Catalog=scooter-rental;Integrated Security=True;"
+  },
   "AllowedHosts": "*"
 }
